requirements.txt
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Pillow==7.0.0
numpy==1.18.1
matplotlib==3.2.1
opencv==3.4.2
